Extract native CLI command construction
Move installer command assembly into a focused builder while preserving the public facade and command behavior. Keep uninstall path resolution in NativeCliInstaller until path responsibilities are separated.
