+++ Packages/src/Editor/Infrastructure/CLI/NativeCliCommandBuilder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffc89cdbedfb948c88b23df88cd41e8c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
